addItem editing
Add item did not give a quantity, so therefore the classes of equipment
and chemical have been used instead. I need to make the addItem method
go to a new menu consisting of equipment and chemical, then run their
input screen(like the one which had item on it) through that.